删除数据集时判断该数据集主从关系

+++ ag-admin/src/main/java/com/yihu/ehr/std/controller/DataSetController.java
@@ -4,13 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yihu.ehr.agModel.standard.datasset.DataSetModel;
 import com.yihu.ehr.agModel.standard.datasset.MetaDataModel;
 import com.yihu.ehr.constants.ApiVersion;
-import com.yihu.ehr.model.standard.MStdDict;
-import com.yihu.ehr.model.standard.MStdSource;
-import com.yihu.ehr.std.service.DataSetClient;
-import com.yihu.ehr.model.standard.MStdDataSet;
-import com.yihu.ehr.model.standard.MStdMetaData;
-import com.yihu.ehr.std.service.DictClient;
-import com.yihu.ehr.std.service.StandardSourceClient;
+import com.yihu.ehr.model.standard.*;
+import com.yihu.ehr.std.service.*;
 import com.yihu.ehr.util.Envelop;
 import com.yihu.ehr.util.controller.BaseController;
 import io.swagger.annotations.ApiOperation;
@@ -39,6 +34,12 @@ public class DataSetController extends BaseController {
 
     @Autowired
     private StandardSourceClient stdSourcrClient;
+
+    @Autowired
+    private CDAVersionClient versionClient;
+
+    @Autowired
+    private CDAClient cdaClient;
 
     @Autowired
     private ObjectMapper objectMapper;
@@ -100,15 +101,43 @@ public class DataSetController extends BaseController {
             @ApiParam(name = "version", value = "版本", defaultValue = "")
             @RequestParam(value = "version") String version) {
 
+        Envelop envelop = new Envelop();
+
         ids = trimEnd(ids, ",");
         if (StringUtils.isEmpty(ids)) {
             return failed("请选择需要删除的数据!");
         }
+
+        envelop = isDeleteDataSet(ids);
+        if (!envelop.isSuccessFlg()){
+            return envelop;
+        }
+
         boolean bo = dataSetClient.deleteDataSet(ids, version);
         if (!bo) {
             return failed("删除失败!");
         }
         return success(null);
+    }
+
+    public Envelop isDeleteDataSet(String id){
+
+        Envelop envelop = new Envelop();
+        envelop.setSuccessFlg(true);
+
+        ResponseEntity<Collection<MCDAVersion>> responseEntity = versionClient.searchCDAVersions("", "", "", 1000, 1);
+        Collection<MCDAVersion> mCdaVersions = responseEntity.getBody();
+
+        for (MCDAVersion mcdaVersion : mCdaVersions) {
+            ResponseEntity<Collection<MCdaDataSetRelationship>> mCdaDataSetRelationships = cdaClient.getCDADataSetRelationships("", "dataSetId=" + id, "", id.length(), 1, mcdaVersion.getVersion());
+            if (mCdaDataSetRelationships.getBody().size()>0){
+                envelop.setSuccessFlg(false);
+                envelop.setErrorMsg("该数据集正被"+mcdaVersion.getVersionName()+"版本，CDA文档使用，不可删除");
+                return envelop;
+            }
+        }
+
+        return envelop;
     }
 
     @RequestMapping(value = "/data_set/{id}", method = RequestMethod.GET)
